omniauth-mailchimp v2.3.0
This upgrades omniauth to 2.x, as the 1.x versions are vulnerable to [CVE-2015-9284][cve]. If you are unable to upgrade,
you can mitigate this vulnerability by either adding CSRF tokens to the routes using one of the methods described in the
bulletin or by forcing the paths to use POST (a good practice anyways).

[cve]: https://github.com/advisories/GHSA-ww4x-rwq6-qpgf

the previous commit/release bumped the version too far in error and did not match the actual version of the gem.
The erroneous gem will be yanked to allow future updates to the 2.x versions.

This is why important work should *not* be done at 1am